Fix Windows CI: add missing cast for GOSUB param_count
The operand assignment at line 402 was missing the (uint32_t) cast
for stmt->as.gosub_stmt.param_count (size_t to uint32_t).

This is the actual source of the C4267 warning on Windows MSVC.

+++ src/c64-script-bytecode.c
@@ -399,7 +399,7 @@ static bool compile_statement(compiler_context_t *ctx, c64script_ast_node_t *stm
     case AST_STMT_GOSUB:
         // Push parameter count first
         emit(ctx, OP_PUSH_NUM, 0, stmt->line);
-        ctx->instructions[ctx->instruction_count - 1].operand = stmt->as.gosub_stmt.param_count;
+        ctx->instructions[ctx->instruction_count - 1].operand = (uint32_t)stmt->as.gosub_stmt.param_count;
 
         // Push each parameter value
         for (size_t i = 0; i < stmt->as.gosub_stmt.param_count; i++) {
